exp/keystore: move keystore out from exp directory (#1431)
This PR also adds a CHANGELOG to the keystore
