+++ README.md
@@ -12,6 +12,7 @@ This repo is the home for all of the public go code produced by SDF.  In additio
 * [Go Horizon SDK - txnbuild](txnbuild): Construct Stellar transactions and operations
 * [Bifrost](services/bifrost): Bitcoin/Ethereum -> Stellar bridge
 * [Ticker](services/ticker): An API server that provides statistics about assets and markets on the Stellar network
+* [Keystore](services/keystore): An API server that is used to store and manage encrypted keys for Stellar client applications
 * Servers for Anchors & Financial Institutions
   * [Bridge Server](services/bridge): send payments and take action when payments are received
   * [Compliance Server](services/compliance): Allows financial institutions to exchange KYC information
