+++ services/keystore/CHANGELOG.md
@@ -1,0 +1,3 @@
+## [v1.0.0] - 2019-06-18
+
+Initial release of the keystore.
