+++ services/keystore/api.go
