+++ services/keystore/api_test.go
