+++ services/keystore/cmd/keystored/main.go
@@ -17,7 +17,7 @@ import (
 
 	migrate "github.com/rubenv/sql-migrate"
 	"github.com/sirupsen/logrus"
-	"github.com/stellar/go/exp/services/keystore"
+	"github.com/stellar/go/services/keystore"
 	"github.com/stellar/go/support/log"
 
 	_ "github.com/lib/pq"
