+++ services/keystore/context.go
