+++ services/keystore/keys_test.go
