+++ services/keystore/spec.md
