+++ services/ticker/CHANGELOG.md
@@ -1,3 +1,3 @@
-## [v1.0.0] - 2018-05-20
+## [v1.0.0] - 2019-05-20
 
 Initial release of the ticker.
